notesapplication/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
identify==2.5.4
idna==3.3
Markdown==3.4.1
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
PyJWT==2.4.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.4
